Move practice slot data to mocks folder
