+++ src/modules/tests/tests.effects.ts
@@ -5,10 +5,9 @@ import { TestPersistenceProvider } from '../../providers/test-persistence/test-p
 import { from } from 'rxjs/observable/from';
 import * as testActions from './tests.actions';
 import { of } from 'rxjs/observable/of';
-import { Candidate } from '@dvsa/mes-test-schema/categories/B';
 import { PopulateApplicationReference } from './application-reference/application-reference.actions';
 import { PopulateCandidateDetails } from './candidate/candidate.actions';
-import { Application } from '../../shared/models/DJournal';
+import { practiceSlot } from './__mocks__/tests.mock';
 
 @Injectable()
 export class TestsEffects {
@@ -56,44 +55,3 @@ export class TestsEffects {
   );
 
 }
-
-export const application: Application = {
-  applicationId: 1234569,
-  bookingSequence: 1,
-  checkDigit: 9,
-  entitlementCheck: false,
-  extendedTest: false,
-  progressiveAccess: false,
-  testCategory: 'B',
-  vehicleGearbox: 'Manual',
-  welshTest: false,
-};
-
-export const candidate: Candidate = {
-  candidateAddress: {
-    addressLine1: 'My House',
-    addressLine2: 'Someplace',
-    addressLine3: 'Sometown',
-    postcode: 'AB45 6CD',
-  },
-  candidateId: 1,
-  candidateName: {
-    firstName: 'Practice',
-    lastName: 'Mode',
-    title: 'Miss',
-  },
-  driverNumber: 'MODEX625220A99HC',
-  mobileTelephone: '07654 123456',
-};
-
-const practiceSlot = {
-  slotDetail: {
-    slotId: 1,
-    duration: 57,
-    start: '2019-01-01T10:14:00+00:00',
-  },
-  booking: {
-    application,
-    candidate,
-  },
-};
